requirements.txt
google-genai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
playwright>=1.40.0
pypdf>=4.0.0
tabulate>=0.9.0
python-docx>=1.1.0
docx2pdf>=0.1.8
docxtpl>=0.16.0
